requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
SQLAlchemy==2.0.38
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
python-dotenv==1.0.1
google-generativeai==0.8.4
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.41
langchain-google==0.1.1
langchain-google-genai==2.0.10
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.6
PyPDF2==3.0.1
faiss-cpu==1.10.0
streamlit==1.43.0
tenacity==9.0.0
sentence-transformers==3.4.1
python-multipart==0.0.20
greenlet==3.1.1
aiosqlite==0.21.0